S3Location moved to absa commons // todo replace snapshot version when commons are released
